requirements.txt

streamlit>=1.22.0
pandas>=1.5.0
matplotlib>=3.6.0
Pillow>=9.3.0
